% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{brew_dyads}
\alias{brew_dyads}
\title{brew_dyads}
\usage{
brew_dyads(.data, .aggregate_date = "year")
}
\arguments{
\item{.data}{passed from `tap_events()` or `connect_phx_db("events")`}

\item{.aggregate_date}{a character string specifying a time unit or a
multiple of a unit to be rounded to. Valid base units are day, week, month,
bimonth, quarter, season, halfyear and year. Default is "year"}
}
\value{
tbl_df with dyadic pairs aggregated by date. In addition, dummy
  variables are generated for all cameo codes requested to capture if that
  cameo coding structure was utilized in that dyad time point.
}
\description{
Construct dyadic data using the `_root`s of source and target variables. All
dyads with missing data for either sideA or sideB are dropped. The data
allows one to quickly aggregated cameo activity between two partners. All
dyadic associations are generated for a comprehensive time range.
}
\examples{

tap_events(cameo_code = c("041"),"SWB") \%>\% 
sample_n(20) \%>\% 
brew_dyads(.aggregate_date = "year")

}
