% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{connect_phx_db}
\alias{connect_phx_db}
\title{connect_phx_db}
\usage{
connect_phx_db(.table = "events")
}
\arguments{
\item{.table}{table being requested. Only two options: "events" or "titles"}
}
\value{
sqlite db connection to the phx database.
}
\description{
Provide a quick and clean connection to phx database to provide specific
individualized queries.
}
\examples{

require(dplyr)
connect_phx_db() \%>\% 
select(date,year)

}
