% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{download_phx_db}
\alias{download_phx_db}
\title{download_phx_db}
\usage{
download_phx_db()
}
\value{
NULL; database installed on local drive.
}
\description{
Download Phoenix Event Database. Database must be downloaded for the
subsequent functions to work.
}
\examples{

download_phx_db()

}
