% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{expand_date}
\alias{expand_date}
\title{expand_date}
\usage{
expand_date(.data, date_variable, .aggregate_date = "year", .fill = 0)
}
\arguments{
\item{.data}{tbl_df with a R-readable date column}

\item{date_variable}{R-readable date column}

\item{.aggregate_date}{level that the temporal unit expansion entries should
be aggregated to}

\item{.fill}{specify what to fill numeric values with; default is 0}
}
\value{
tbl_df with the date gaps filled
}
\description{
Expand a date range
}
\examples{

require(dplyr)
connect_phx_db() \%>\% 
filter(code=="041") \%>\% 
select(date,code) \%>\% 
head(30) \%>\% 
collect() \%>\% 
expand_date(date,.aggregate_date = "day",.fill=0) 

}
