% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{map_meta}
\alias{map_meta}
\title{map_meta}
\usage{
map_meta(.data)
}
\arguments{
\item{.data}{output from `tap_events()`}
}
\value{
return tbl_df with mapped title information on it.
}
\description{
Map meta data for onto event entries assuming that output follows from
`tap_events()`.
}
\examples{

tap_events(cameo_code = "041",.database = "NYT") \%>\% 
map_meta()

}
