% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{tap_codebook}
\alias{tap_codebook}
\title{tap_codebook}
\usage{
tap_codebook(output_html = T)
}
\arguments{
\item{output_html}{codebook should be opened in the browser, else will be
opened as a PDF using the system drive; default is TRUE.}
}
\value{
Opens the current Phoenix Codebook
}
\description{
Open the Current Phoenix Codebook.
}
\examples{

tap_codebook(output_html=F) # for PDF

}
