% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{tap_events}
\alias{tap_events}
\title{tap_events}
\usage{
tap_events(cameo_code = "all", .database = "all")
}
\arguments{
\item{cameo_code}{select the specific cameo code(s). Cameo codes must be
presented as strings, e.g. "041". If more than one code is select, do so as
a vector using `c()`}

\item{.database}{select database (FBIS, NYT, SWB); default is that all
databases are selected from}
}
\value{
tbl_df of requested cameo coded events.
}
\description{
extract events from the Phoenix Event Dataset.
}
\examples{

tap_events(cameo_code = c("041"),"SWB")

}
