% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{tap_meta}
\alias{tap_meta}
\title{tap_meta}
\usage{
tap_meta(.database = "all")
}
\arguments{
\item{.database}{select database (FBIS, NYT, SWB); default is that all
databases are selected from}
}
\value{
tbl_df of all data.
}
\description{
Extract event meta data from the Phoenix Event Dataset.
}
\examples{

tap_meta(.database="NYT") \%>\% head()

}
