% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{taste_events}
\alias{taste_events}
\title{taste_events}
\usage{
taste_events(.database = "all")
}
\arguments{
\item{.database}{select database (FBIS, NYT, SWB); default is that all
databases are selected from}
}
\description{
Preview event entries from the Phoenix Event Dataset.
}
\examples{

taste_events(database="SWB")
}
