% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{taste_meta}
\alias{taste_meta}
\title{taste_meta}
\usage{
taste_meta(.database = "all")
}
\arguments{
\item{.database}{select database (FBIS, NYT, SWB); default is that all
databases are selected from}
}
\value{
return a sql preview of the available data.
}
\description{
Preview meta data for event entries from the Phoenix Event Dataset.
}
\examples{

taste_meta(.database="NYT")

}
