% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{translate_actors}
\alias{translate_actors}
\title{translate_actors}
\usage{
translate_actors(.actor_code = "")
}
\arguments{
\item{.actor_code}{actor codes for the Phoenix Event Data.}
}
\value{
vector of translated actor names as they appear in the Cline codebook.
}
\description{
Provides a key for all actor entries in the Phoenix event data. Data drawn
from the Cline code book.
}
\examples{

translate_actors(c("SUN","USA","SUN","RUS","REB"))

}
