% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{translate_cameo}
\alias{translate_cameo}
\title{translate_cameo}
\usage{
translate_cameo(.cameo_code = "all")
}
\arguments{
\item{.cameo_code}{select the specific cameo code(s). Cameo codes must be
presented as strings, e.g. "041". If more than one code is select, do so as
a vector using `c()`}
}
\value{
tbl_df of cameo code (as character) and cameo txt
}
\description{
Provide a key for all cameo codes into plain text to make translation clear.
}
\examples{

translate_cameo(c("041","032","062"))

}
