% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{download_phx_db}
\alias{download_phx_db}
\title{download_phx_db}
\usage{
download_phx_db(manual_path = NULL)
}
\arguments{
\item{manual_path}{{temporary} character string of file path of curated
database to be uploaded into package infrastructure. Default NULL.}
}
\value{
NULL; database installed on local drive.
}
\description{
Download Phoenix Event Database. Database must be downloaded for the
subsequent functions to work.
}
\examples{

download_phx_db()

}
